add user roles
This adds the concept of roles to users so that different permissions
can be granted based on those roles. Additionally it locks down the live
dashboard so that it can only be accessed by those with an admin role.

Also this includes style updates to the various authentication pages to
prepare for allowing users to create accounts.