add recipe list endpoint
Rather than have the mobile app have to traverse each of the categories
in order to fetch all of the recipes, it's easier to add a single
endpoint from which the app can download all of the data. Additionally
the endpoint takes in a parameter for the last sync date so that only
recipes that have been updated since that date will be fetched.