use category_slug instead of category_id
When users were only adding recipes via the web UI, it was fairly simple
to use the numeric category ID for associating a recipe with a category.
The API doesn't surface the numeric category ID anywhere though, so it's
impossible for an API user to have the category's numeric ID to be able
to create a recipe with.

This adds a new category_slug field for recipes so that the slug can be
given via the API for creating a new recipe. It also updates the UI and
code to use the new field instead of the category_id field. Once this is
in place, the category_slug field will need to be backfilled for
everything to work correctly. Then the field will need to be updated to
have a NOT NULL constraint, and the category_id column can be deleted
from the database.